lake-to-warehouse/requirements.txt
Flask>1
slack-bolt==1.4.4
PyYAML==5.4.1
yapf==0.31.0
pytz==2021.1
pytest==6.2.3
google-cloud-storage==1.37.1
jsonlines==2.0.0
google-cloud-pubsub
google-cloud-bigquery
